Fix behaviour for blocking used channels and add type-check to channel equality in test cases.
