+++ tests/wirer/test_add_dummy_line.py
@@ -1,5 +1,3 @@
-from dataclasses import asdict
-
 import pytest
 from qualang_tools.wirer import Connectivity, allocate_wiring, visualize
 from qualang_tools.wirer.connectivity.element import Element, Reference
@@ -30,7 +28,7 @@ def test_add_dummy_line(instruments_2lf_2mw):
     # regression test
     test_element = connectivity.elements[Reference('test')]
     for i, channel in enumerate(test_element.channels['ch']):
-        assert asdict(channel) == asdict([
+        assert pytest.channels_are_equal(channel, [
             InstrumentChannelLfFemInput(con=1, port=1, slot=1),
             InstrumentChannelLfFemOutput(con=1, port=6, slot=1)
         ][i])
