+++ tests/wirer/test_wirer_lf_and_mw.py
@@ -1,5 +1,3 @@
-from dataclasses import asdict
-
 import pytest
 
 from qualang_tools.wirer import *
@@ -27,30 +25,30 @@ def test_6q_allocation(instruments_2lf_2mw):
 
     for qubit in qubits:
         # flux channels should have some port as qubit index since they're allocated sequentially
-        flux_channels = connectivity.elements[QubitReference(qubit)].channels[WiringLineType.FLUX]
-        assert [asdict(ch) for ch in flux_channels] == [
-            asdict(InstrumentChannelLfFemOutput(con=1, port=qubit, slot=1))
-        ]
+        for i, channel in enumerate(connectivity.elements[QubitReference(qubit)].channels[WiringLineType.FLUX]):
+            assert pytest.channels_are_equal(channel, [
+                InstrumentChannelLfFemOutput(con=1, port=qubit, slot=1)
+            ][i])
 
         # resonators all on same feedline, so should be first available input + outputs channels on MW-FEM
-        resonator_channels = connectivity.elements[QubitReference(qubit)].channels[WiringLineType.RESONATOR]
-        assert [asdict(ch) for ch in resonator_channels] == [
-            asdict(InstrumentChannelMwFemInput(con=1, port=1, slot=3)),
-            asdict(InstrumentChannelMwFemOutput(con=1, port=1, slot=3))
-        ]
+        for i, channel in enumerate(connectivity.elements[QubitReference(qubit)].channels[WiringLineType.RESONATOR]):
+            assert pytest.channels_are_equal(channel, [
+                InstrumentChannelMwFemInput(con=1, port=1, slot=3),
+                InstrumentChannelMwFemOutput(con=1, port=1, slot=3)
+            ][i])
 
         # drive channels are on MW-FEM
-        drive_channels = connectivity.elements[QubitReference(qubit)].channels[WiringLineType.DRIVE]
-        assert [asdict(ch) for ch in drive_channels] == [
-            asdict(InstrumentChannelMwFemOutput(con=1, port=qubit+1, slot=3))
-        ]
+        for i, channel in enumerate(connectivity.elements[QubitReference(qubit)].channels[WiringLineType.DRIVE]):
+            assert pytest.channels_are_equal(channel, [
+                InstrumentChannelMwFemOutput(con=1, port=qubit+1, slot=3)
+            ][i])
 
     for i, pair in enumerate(qubit_pairs):
         # coupler channels should have some port as pair index since they're allocated sequentially, but on slot 2
-        coupler_channels = connectivity.elements[QubitPairReference(*pair)].channels[WiringLineType.COUPLER]
-        assert [asdict(ch) for ch in coupler_channels] == [
-            asdict(InstrumentChannelLfFemOutput(con=1, port=i+1, slot=2))
-        ]
+        for j, channel in enumerate(connectivity.elements[QubitPairReference(*pair)].channels[WiringLineType.COUPLER]):
+            assert pytest.channels_are_equal(channel, [
+                InstrumentChannelLfFemOutput(con=1, port=i+1, slot=2)
+            ][j])
 
 def test_4rr_allocation(instruments_2lf_2mw):
     connectivity = Connectivity()
@@ -67,9 +65,9 @@ def test_4rr_allocation(instruments_2lf_2mw):
 
     # resonators all on different feedlines, so should fill all 4 inputs of 2x MW-FEM
     for i, qubit in enumerate([1, 2, 3, 4]):
-        resonator_channels = connectivity.elements[QubitReference(qubit)].channels[WiringLineType.RESONATOR]
-        assert [asdict(ch) for ch in resonator_channels] == [
-            asdict(InstrumentChannelMwFemInput(con=1, port=[1, 2, 1, 2][i], slot=[3, 3, 7, 7][i])),
-            asdict(InstrumentChannelMwFemOutput(con=1, port=[1, 2, 1, 2][i], slot=[3, 3, 7, 7][i]))
-        ]
+        for j, channel in enumerate(connectivity.elements[QubitReference(qubit)].channels[WiringLineType.RESONATOR]):
+            assert pytest.channels_are_equal(channel, [
+                InstrumentChannelMwFemInput(con=1, port=[1, 2, 1, 2][i], slot=[3, 3, 7, 7][i]),
+                InstrumentChannelMwFemOutput(con=1, port=[1, 2, 1, 2][i], slot=[3, 3, 7, 7][i])
+            ][j])
